--- a/IntelligentCache.Tests/FakeRedis.cs
+++ b/IntelligentCache.Tests/FakeRedis.cs
@@ -1,3 +1,5 @@
+#pragma warning disable RCS1102 // Make class static.
+
 using FakeItEasy;
 using StackExchange.Redis;
 using System;
@@ -7,7 +9,7 @@ namespace IntelligentCache.Tests
 {
     internal class FakeRedis
     {
-        public static IConnectionMultiplexer CreateConnectionMultiplexer(Func<RedisKey, RedisValue>? onGet = null, Action<RedisKey, RedisValue, TimeSpan?>? onSet = null)
+        public static IConnectionMultiplexer CreateConnectionMultiplexer(Func<RedisKey, RedisValue> onGet = null, Action<RedisKey, RedisValue, TimeSpan?> onSet = null)
         {
             var multiplexer = A.Fake<IConnectionMultiplexer>(o => o.Strict());
             var database = A.Fake<IDatabase>(o => o.Strict());
@@ -15,20 +17,20 @@ namespace IntelligentCache.Tests
             A.CallTo(() => multiplexer.GetDatabase(A<int>._, A<object>._)).Returns(database);
 
             A.CallTo(() => database.StringGet(A<RedisKey>._, A<CommandFlags>._))
-                .ReturnsLazily((RedisKey key, CommandFlags flags) => onGet?.Invoke(key) ?? RedisValue.Null);
+                .ReturnsLazily((RedisKey key, CommandFlags _) => onGet?.Invoke(key) ?? RedisValue.Null);
 
             A.CallTo(() => database.StringGetAsync(A<RedisKey>._, A<CommandFlags>._))
-                .ReturnsLazily((RedisKey key, CommandFlags flags) => onGet?.Invoke(key) ?? RedisValue.Null);
+                .ReturnsLazily((RedisKey key, CommandFlags _) => onGet?.Invoke(key) ?? RedisValue.Null);
 
             A.CallTo(() => database.StringSet(A<RedisKey>._, A<RedisValue>._, A<TimeSpan?>._, A<When>._, A<CommandFlags>._))
-                .ReturnsLazily((RedisKey key, RedisValue value, TimeSpan? expiry, When when, CommandFlags flags) =>
+                .ReturnsLazily((RedisKey key, RedisValue value, TimeSpan? expiry, When _, CommandFlags __) =>
                 {
                     onSet?.Invoke(key, value, expiry);
                     return true;
                 });
 
             A.CallTo(() => database.StringSetAsync(A<RedisKey>._, A<RedisValue>._, A<TimeSpan?>._, A<When>._, A<CommandFlags>._))
-                .ReturnsLazily((RedisKey key, RedisValue value, TimeSpan? expiry, When when, CommandFlags flags) =>
+                .ReturnsLazily((RedisKey key, RedisValue value, TimeSpan? expiry, When _, CommandFlags __) =>
                 {
                     onSet?.Invoke(key, value, expiry);
                     return true;
@@ -37,23 +39,20 @@ namespace IntelligentCache.Tests
             return multiplexer;
         }
 
-        public static ISubscriber CreateSubscriber(Action<RedisChannel, RedisValue>? onPublish = null)
+        public static ISubscriber CreateSubscriber(Action<RedisChannel, RedisValue> onPublish = null)
         {
             var subscriber = A.Fake<ISubscriber>();
 
             var subscriptions = new List<Action<RedisChannel, RedisValue>>();
 
             A.CallTo(() => subscriber.Subscribe(A<RedisChannel>._, A<Action<RedisChannel, RedisValue>>._, A<CommandFlags>._))
-                .Invokes((RedisChannel channel, Action<RedisChannel, RedisValue> handler, CommandFlags flags) =>
-                {
-                    subscriptions.Add(handler);
-                });
+                .Invokes((RedisChannel _, Action<RedisChannel, RedisValue> handler, CommandFlags __) => subscriptions.Add(handler));
 
             A.CallTo(() => subscriber.Publish(A<RedisChannel>._, A<RedisValue>._, A<CommandFlags>._))
-                .Invokes((RedisChannel channel, RedisValue message, CommandFlags flags) => PublishHandler(channel, message));
+                .Invokes((RedisChannel channel, RedisValue message, CommandFlags _) => PublishHandler(channel, message));
 
             A.CallTo(() => subscriber.PublishAsync(A<RedisChannel>._, A<RedisValue>._, A<CommandFlags>._))
-                .Invokes((RedisChannel channel, RedisValue message, CommandFlags flags) => PublishHandler(channel, message));
+                .Invokes((RedisChannel channel, RedisValue message, CommandFlags _) => PublishHandler(channel, message));
 
             void PublishHandler(RedisChannel channel, RedisValue message)
             {