--- a/IntelligentCache.Tests/IntelligentCache.Tests.csproj
+++ b/IntelligentCache.Tests/IntelligentCache.Tests.csproj
@@ -4,12 +4,11 @@
         <TargetFrameworks>net461;netcoreapp3.1;net48;net5.0</TargetFrameworks>
         <IsPackable>false</IsPackable>
         <LangVersion>8.0</LangVersion>
-        <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
         <PackageReference Include="FakeItEasy" Version="7.0.2" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.10.0" />
         <PackageReference Include="xunit" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
           <PrivateAssets>all</PrivateAssets>