--- a/IntelligentCache.sln
+++ b/IntelligentCache.sln
@@ -11,7 +11,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntelligentCache.Tests", "IntelligentCache.Tests\IntelligentCache.Tests.csproj", "{DE5B0C27-C362-4EDB-8C88-8973B5E7B3CA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntelligentCache.Tests", "IntelligentCache.Tests\IntelligentCache.Tests.csproj", "{DE5B0C27-C362-4EDB-8C88-8973B5E7B3CA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution