--- a/IntelligentCache/MemoryCache.cs
+++ b/IntelligentCache/MemoryCache.cs
@@ -6,15 +6,18 @@ using MemCache = System.Runtime.Caching.MemoryCache;
 namespace IntelligentHack.IntelligentCache
 {
     /// <summary>
-    /// An implementation of <see cref="ICache" /> that stores values in memory.
+    /// An implementation of <see cref="ICache" /> that stores values in a <see cref="System.Runtime.Caching.MemoryCache"/>.
     /// </summary>
     public class MemoryCache : ICache
     {
         private readonly string _prefix;
         private readonly object _synclock = new object();
 
-        /// <param name="prefix">A prefix that is inserted before each key to prevent collisions with other users of the shared cache.</param>
         public MemoryCache(string prefix)
+        /// <summary>
+        /// Creates a cache that runs in the server memory.
+        /// </summary>
+        /// <param name="prefix">This string is prefixed to the key names to partition the keys if the underlying storage is shared</param>
         {
             if (prefix is null)
             {
@@ -24,6 +27,7 @@ namespace IntelligentHack.IntelligentCache
             _prefix = prefix + ":";
         }
 
+        /// <inheritdoc />
         public T GetSet<T>(string key, Func<T> calculateValue, TimeSpan duration) where T : class
         {
             var k = _prefix + key;
@@ -57,12 +61,14 @@ namespace IntelligentHack.IntelligentCache
             MemCache.Default.Remove(k);
         }
 
+        /// <inheritdoc />
         public Task<T> GetSetAsync<T>(string key, Func<CancellationToken, Task<T>> calculateValue, TimeSpan duration, CancellationToken cancellationToken = default) where T : class
         {
             var result = GetSet(key, () => calculateValue(cancellationToken).GetAwaiter().GetResult(), duration);
             return Task.FromResult(result);
         }
 
+        /// <inheritdoc />
         public Task InvalidateAsync(string key, CancellationToken cancellationToken = default)
         {
             Invalidate(key);