--- a/IntelligentCache/PassThroughCache.cs
+++ b/IntelligentCache/PassThroughCache.cs
@@ -13,21 +13,25 @@ namespace IntelligentHack.IntelligentCache
     /// </remarks>
     public sealed class PassThroughCache : ICache
     {
+        /// <inheritdoc />
         public T GetSet<T>(string key, Func<T> calculateValue, TimeSpan duration) where T : class
         {
             return calculateValue();
         }
 
+        /// <inheritdoc />
         public Task<T> GetSetAsync<T>(string key, Func<CancellationToken, Task<T>> calculateValue, TimeSpan duration, CancellationToken cancellationToken = default) where T : class
         {
             return calculateValue(cancellationToken);
         }
 
+        /// <inheritdoc />
         public void Invalidate(string key)
         {
             return;
         }
 
+        /// <inheritdoc />
         public Task InvalidateAsync(string key, CancellationToken cancellationToken = default)
         {
             return Task.CompletedTask;