@ocoster-is 2021-06-07
The only concern I have with this is that the dictionary in `MultiKeyLock` is unbounded and there is no way to evict keys from it. Over time this dictionary could become a bottleneck if it gains a lot of entries (and can become a memory hog, as each `ReaderWriterLockSlim` takes > 40 bytes).

Mentioning this as this class is used in long lived applications (the caches).