@aaubry 2021-06-07
Good catch, it should probably be nuked on invalidate, at the very least.
Perhaps we could... use a cache for it? /ducks

On Mon, Jun 7, 2021 at 1:23 PM Oded Coster ***@***.***> wrote:

> The only concern I have with this is that the dictionary in MultiKeyLock
> is unbounded and there is no way to evict keys from it. Over time this
> dictionary could become a bottleneck if it gains a lot of entries (and can
> become a memory hog, as each ReaderWriterLockSlim takes > 40 bytes).
>
> Mentioning this as this class is used in long lived applications (the
> caches).
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/intelligenthack/intelligentcache/pull/33#issuecomment-855843075>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAEZ5X7SQZ6GDZLFU4VZMB3TRST3TANCNFSM46EX42XQ>
> .
>


-- 
Marco Cecconi

CEO |Intelligent Hack

+39 3496356763

www.intelligenthack.com <https://intelligenthack.com/en>

<https://twitter.com/intelligenthack>
<https://www.facebook.com/intelligenthack>
<https://www.linkedin.com/company/intelligent-hack>